+++ docs/CHANGE_LOG.md
@@ -1,5 +1,13 @@
 # Syntax.js - Change Log:
 
+## Version 0.8.0:
+- Added a new setting called "showLineNumbers", which states if the line numbers should be shown (defaults to true).
+- Renamed the public function "buildNewSyntaxElements" to "findAndBuildNewElements()".
+- Added new public function "destroy()", which will revert all rendered elements to their original HTML (minus the syntax attributes).
+
+<br>
+
+
 ## Version 0.7.0:
 - Added NuGet.org Nuget package support via a new nuspec file.
 - The default language "JavaScript" is no longer built-in, and must be imported like all other languages.